DOMAIN Y CONTROLLER creados y funcionando --- Pendiente -> (UseCases)
