+++ internal/usecase/item_usecase.go
@@ -1,0 +1,1 @@
+package usecase //Aplicacion Bussiness Rules
